move dashboardbasemixin to mixins.py
